AVRO-1568 add customEncoder that adds polymorphic support
loosely based on the jackson approach. But not given the option to
specify a name. On the interface or superclass add the custom encoder.

If using java >=17 can use sealed interfaces and permits to identify
available types.

Alternatively can use an annotation to define the sub types.

Creates a schema that is a union of all defined / permitted sub types